Tests: reworked try_run() to redirect stderr.
Previously, if TEST_NGINX_VERBOSE was set, the code tried to use compiled-in
error log.  While this works in typical development cases when --error-log-path
is relative to prefix, it is unlikely to work correctly if --error-log-path
is absolute.

The recently introduced '-e' command line option might be used to fix this.
On the other hand, a more compatible solution would be to redirect stderr
to a file instead.

--HG--
branch : nginx-tests