Removed extra allocation for r->uri.
The ngx_http_parse_complex_uri() function cannot make URI longer and does
not null-terminate URI, so there is no need to allocate an extra byte.  This
allocation appears to be a leftover from changes in 461:7bdcc2986461 (0.1.5),
where null-termination of r->uri and many other strings was removed.

--HG--
branch : nginx