Mail: POP3 pipelining support.
The change is mostly the same as the SMTP one (9da69a037164 and adb81f63edc4),
and ensures that nginx is able to properly handle or reject multiple POP3
commands, as required by the PIPELINING capability (RFC 2449).  The s->cmd
field is not really used and set for consistency.

--HG--
branch : nginx