Mail: IMAP pipelining support.
The change is mostly the same as the SMTP one (9da69a037164 and adb81f63edc4),
and ensures that nginx is able to properly handle or reject multiple IMAP
commands.  The s->cmd field is not really used and set for consistency.

Non-synchronizing literals handling in invalid/unknown commands is limited,
so when a non-synchronizing literal is detected at the end of a discarded
line, the connection is closed.

--HG--
branch : nginx