Tests: mirror_proxy.t speedup.
The failing "mirror delay" test doesn't have to wait for a hung connection.
The hang is anyway manifested with a socket leak alert.  This eliminates an
unnecessary read timeout delay in the client and allows to enable the test
on win32.  Additionally, in the mirror_request_body test, disabled passing
Content-Length to backend to avoid wasting time waiting for an absent body.

--HG--
branch : nginx-tests