Use only preallocated memory in ngx_readv_chain() (ticket #1408).
In dac99c630306, the number of preallocated iovec's for ngx_readv_chain()
was increased.  Still, in some setups, the function might allocate memory
for iovec's from a connection pool, which is only freed when closing the
connection.

The ngx_readv_chain() function was modified to use only preallocated
memory, similarly to the ngx_writev_chain() change in 21388d58f946.

--HG--
branch : nginx