Generic HTTP API module and underline logic.
This adds the "api" directive, which can be specified in a location context,
making particular location as an entry point for HTTP API requests.

It uses JSON format for data representation and fully traversable via URI.
E.g. if an API tree has the following JSON structure:

{
    "foo": {
        "bar": 42
    }
}

then a GET request to "/" will return the object above as is, while a request
to "/foo/" will return:

{
    "bar": 42
}

and finally, a request to "/foo/bar/" will return just 42.

Note that ending slash doesn't matter.

Currently this API consists only of one section "/status/", that supposed to
accumulate all the status information and statistic metrics.

By default, it renders formatted JSON with identation, that is easy for reading
by humans.  In order to save some traffic and avoid extra spaces you can sepcify
"?pretty=off" in arguments.

In case of error, a JSON object with the following fields is returned:

 - "error" (string): a short identificator of the error, currently one
   of the following: "PathNotFound", "MethodNotAllowed", "BadRequest",
   or "InternalError".

 - "description" (string): a human readable description of the error

and HTTP response code is set respectively.