API: handling for case when URI is rewritten inside location.
Previously, the beginning of request URI was truncated by the length of location
prefix assuming that URI matches to that location.  But URI could be rewritten
inside location making that assumption no longer true.  Even more, URI could
get even shorter, resulting in segmentation fault in configuration like this:

  location /api/ {
      rewrite ^ / break;
      api;
  }

Now the truncation is only made if URI either hasn't been changed inside the
location, or the beggining of URI actually matches to the location.