API: relaxed path handling.
This change adds toleration to duplicate slashes and to missing one
in the beginning of the API request path.

It allows for API path to work independently of "merge_slashes" setting and
also handle cases when it constructed by user out of different parts without
caring about slash duplication, which is especially important for adding
support of regexp locations in the following changes.